requirements.txt
fastapi
uvicorn
neo4j
python-dotenv
pandas
pydantic
requests
langchain
langchain-neo4j
langchain-groq
torch
torch-geometric
networkx
scikit-learn
numpy
streamlit
plotly
folium
streamlit_folium